+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## NEXT
 
-## v0.2.0 (22Oct2025)
+## v0.3.0
+
+- Add `NumpyMixin` class
+- Append `Mixin` to remaining mixin classes
+
+## v0.2.0
 
 - Full working implementation with tests.
 
-## v0.1.0 (10Mar2025)
+## v0.1.0
 
 - Hello world!